Skybox: Refactored EquirectangularToCube process.
